refactor: (#108) 상수 파일 통합
