--- src/outfit/outfit.constants.ts
@@ -1,13 +1,0 @@
-import { OutfitIds } from './outfit.type';
-
-export const DEFAULT_BIRD_ID = 150;
-
-export const DEFAULT_OUTFIT: OutfitIds = {
-  bird: 150,
-  hat: null,
-  accessory: null,
-  top: null,
-  bottom: null,
-  shoes: null,
-  effect: null,
-};
